throws error if deletion attempt on nonexistent id, commented ctlr and model

+++ models/products/PaymentOptionsModel.js
@@ -3,6 +3,7 @@
 const sqlite3 = require('sqlite3');
 const db = new sqlite3.Database('./db/api-sprint.sqlite');
 
+// get single payment option by id
 module.exports.getSinglePaymentOption = id => {
   return new Promise((resolve, reject) => {
     db.all(`SELECT * FROM Payment_Options
@@ -13,6 +14,7 @@ module.exports.getSinglePaymentOption = id => {
   });
 };
 
+// get all payment options
 module.exports.getAllPaymentOptions = () => {
   return new Promise((resolve, reject) => {
     db.all(`SELECT * FROM Payment_Options`, (err, data) => {
@@ -22,6 +24,7 @@ module.exports.getAllPaymentOptions = () => {
   });
 };
 
+// create new payment option
 module.exports.createPaymentOption = ({type, account_number, customer_id}) => {
   return new Promise((resolve, reject) => {
     db.run(`INSERT INTO Payment_Options (
@@ -39,6 +42,7 @@ module.exports.createPaymentOption = ({type, account_number, customer_id}) => {
   });
 };
 
+// update one payment option by id
 module.exports.updatePaymentOption = (id, {type, account_number, customer_id}) => {
   return new Promise((resolve, reject) => {
     db.run(`REPLACE INTO Payment_Options (
@@ -58,11 +62,20 @@ module.exports.updatePaymentOption = (id, {type, account_number, customer_id}) =
   });
 };
 
+// delete one payment option by id
 module.exports.deletePaymentOption = id => {
   return new Promise((resolve, reject) => {
-    db.run(`DELETE FROM Payment_Options WHERE payment_option_id = ${id}`, err => {
-      if (err) return reject(err);
-      resolve(id);
-    });
+    module.exports.getSinglePaymentOption(id)
+      .then(data => {
+        if (data.length >= 1) {
+          db.run(`DELETE FROM Payment_Options WHERE payment_option_id = ${id}`, err => {
+            if (err) return reject(err);
+            resolve(id);
+          });
+        } else {
+          let err = new Error("There is no Payment Option with that ID.");
+          reject(err);
+        }
+      })
   });
 };
